D: `seznam.cz##div[style*="max-height:"]`
Hi, @tomasko126, I am absolutely sorry but there seems to be a change in the way the content is displaed at `https://www.seznam.cz/` and this filter supposed to hide the side banner is crushing the page styling. Could you please remove it, we are looking into finding another ways of getting rid of that ad, without breaking the style. Thank a lot.